feat: implement contact form schema and integrate image handling in IncidentReport
